Split StatsListener into abstract BaseStatsListener + concrete StatsListener (for different types)
